requirements.txt
lxml
python-decouple
apache-airflow-providers-apache-spark
requests
pandas
pg

sqlalchemy 
# psycopg2
tqdm
urllib3 
xlrd
geopandas
pyproj
openpyxl